+++ sever/src/main/java/com/example/rememberdokdo/Repository/Inventory/InventoryItemsRepository.java
@@ -1,6 +1,6 @@
-package com.example.rememberdokdo.Repository;
+package com.example.rememberdokdo.Repository.Inventory;
 
-import com.example.rememberdokdo.Entity.InventoryItemsEntity;
+import com.example.rememberdokdo.Entity.Inventory.InventoryItemsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
